Bug fixes with the way loggers are stored internally for sessions
Loggers now are stored with the key of the arguments that defined them.
This means there will be new loggers created for every new timestamp,
dataset name, tags, etc.

Not 100% sure if this is what we want for this API so I'll have to check
with everyone else. It is definitely better then what's there but it may
not be right.