Added example for anonymous sessions with dates
This adds a snapshot of a dataset from kaggle
(https://www.kaggle.com/yugagrawal95/sample-media-spends-data) and an
example that shows how properly log that data to whylogs for an
anonymous session on WhyLabs. This dataset actually has a day date
column so its much clearer how the logger is supposed to work in regards
to dates.