Add generic retry utility and make profile uploads retry
A lot of resources online point towards leaning on akka to handle things
like this but it seems overkill as a dependnecy for this use case. The
one thing I need to confirm before pushing is the relationship between
futures and threads in scala. In Kotlin, I would just be using a
coroutine specific delay that I know won't deadlock. Here, if its true
that every future runs on its own thread then this should be fine. If
that isn't true then it can be risky.