Fix reversed frequent item upper/lower bounds (#1013)
## Description

It should be the case for `FrequentItem`s that lower bound <= estimate
<= upper bound. It looks like we got the upper & lower bounds reversed
between the pybind11 interface code and the `FrequentItemsMetric`

`items = [FrequentItem(value=x[0], est=x[1], upper=x[2], lower=x[3]) for
x in limited_freq_items]` in `metrics.py` but
```
    py::tuple t = py::make_tuple(iter->get_item(),
                                 iter->get_estimate(),
                                 iter->get_lower_bound(),
                                 iter->get_upper_bound());
```
in `fi_wrapper.cpp` (in the `datasketches-cpp` repo).

## Changes

-Upper/lower bounds fixed

## Related



<!-- Reference related commits, issues and pull requests. Type `#` and
select from the list. -->


<!-- List issues for GitHub to automatically close after merge to
default branch. Type `#` and select from the list. See the [GitHub docs
on linking PRs to
issues](https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue)
for more. -->

- [ ] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).

Co-authored-by: Jamie Broomall <88007022+jamie256@users.noreply.github.com>
Co-authored-by: Jamie Broomall <jamie@whylabs.ai>