Adding Fastapi example (#1053)
## Description

Added a FastAPI example to integrations folder.

The commits in this pull request will <!-- Summarize PR. -->

## Changes

- Imperative commit or change title (commit ID) <!-- Add commit ID and
[GitHub will
autolink](https://help.github.com/en/github/writing-on-github/autolinked-references-and-urls).
-->
  
- two .py files for FastAPI example
- a Readme explaining example

## Related

Relates to organization/repo#number

<!-- Reference related commits, issues and pull requests. Type `#` and
select from the list. -->  

Closes organization/repo#number

<!-- List issues for GitHub to automatically close after merge to
default branch. Type `#` and select from the list. See the [GitHub docs
on linking PRs to
issues](https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue)
for more. -->

- [ ] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).

Co-authored-by: murilommen <murilommen@gmail.com>
Co-authored-by: Jamie Broomall <88007022+jamie256@users.noreply.github.com>