Add a new rolling logger type that handles historical data
The current implementation of rolling logger has a few properties that
make it unusable for the python whylogs container:

- It assumes all logged data has a dataset timestamp of now() and merges
  it into the current (only) profile.
- It has potential to upload data for future dates.
- It conflates upload schedule with dataset profile timestamp.
- It has the potential to tear logging between dataset profiles if
  logging happens during writing.

The new logger breaks the interface of the current Logger to enable
different behavior. Now, the log() method takes in a timestamp that
indicates when the data being logged happened, which is used to
determine what dataset profile or segment cache the data should be
logged to. Internally, it maintains a map of dataset timestamp to
dataset profile/segment cache. The dataset timestamp is created by the
aggregate_by configuration, where it can be Hourly, Daily, etc. For
Hourly, the timestamp will be truncated to the hour.

Dataset timestamps are always the floor of the time interval so no data
can be uploaded to future times.

You can separately configure the upload interval and the model type,
which is important for the container since it has to distinguish between
the two in the context of late data.

The tearing issue is resolved by using a message queue internally for
all actions so that they're handled in serial. This does mean that
logging is paused while uploading happens but log() is also asynchronous
by default now so that won't add additional latency to the caller. The
message queue is processed using a Thread. It's up to the caller to
close the logger during application shutdown as well, which will start a
synchronous series of cleanup events and profile writing.