Multi-column constraint proposal (#1033)
Refactor to support multi-column constraints

Adds `DatasetConstraint` and some other refactoring

A `DatasetConstraint` implements constraints that aren't attached to a
specific column. It has a `Callable` that takes a `DatasetProfileView`.
It returns a `bool` and a dictionary mapping `column_name/metric_name`
to the metrics used in evaluating the constraint. The latter is
necessary for generating the report.

It looks like there's a similar case of a `MetricConstraint` that
applies to all metrics of a specified type across all columns. `MetricConstraintWrapper` turns such a `MetricConstraint` into the
`Callable` for an equivalent `DatasetConstraint`.

I implemented a `PrefixCondition` that implements the prefix
expression syntax used to serialize the `Predicate` expressions with the
addition of a few arithmetic operators. This is probably a reasonable
serialization format, but might not be the ultimate user interface for
specifying constraints...