Update improve README.md file (#1146)
## Description

This repository needs polishing the text of README.md file.

The commits in this pull request will improve the text, thus simplifying
the understanding of anyone reading the file.

## Changes

- Update improved README.md file (d4c692b) 
- Duplicated word "about about" was replaced by "about what a
profile..." resulting in a more meaningful sentence.
- Unnecessary use of words made sentence more complicated than needed.

In line 251 of README.md file the word "here" seems to refer to a link
yet to be implemented. I found no section specifically about whylogs
scaling in the docs so I couldn't include it, but if you give me the
reference I can add the change for this pull request as well.


- [✓] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).