pyspark embeddings support on vectors (#1149)
## Description

the pyspark mapInPandas does not support dataframe that contain vectors,
and the whylogs pyspark profiler uses mapInPandas. To support embeddings
which might be passed in as DenseVectors we check for any columns of
vector type and convert these to arrays inside the whylogs
`collect_dataset_profile_view` if present. This creates a transformed
dataframe which conforms to the requirements of mapInPandas and is
consistent with how whylogs profiles embeddings (counted as tensor
types).

As a side effect we now can profile dataframes that contain vector
columns, where before this change a vector column would cause an error
in the whylogs profiling when using pyspark (even if the configuration
did not include embeddings metrics).

## Changes

- Check for VectorUDT columns inside pyspark profiling functions
- transform vectors to arrays before passing to mapInPandas inside
whylogs pyspark profiling.
- added tests for vectors and embeddings metrics with pyspark.

- [x] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).