Add new apis to enable guest sessions from whylogs
This change introduces the concept of a session that can be initialized
in order to enable a new UX that integrates with WhyLabs, without
requiring any initial user sign up. The new apis work as follows

```python
import whylogs as why

why.init(session_type='whylabs_anonymous')

result_set = why.log(df, name="foo")

result_set = why.log(multiple={"foo": df1, "bar": df2})
```

The `init` api is optional. Existing users can continue to do what
they're doing today. It can be called with
`session_type='whylabs_anonymous'` to create an anonymous session. A
session is mostly an internal detail. It's something that is available
to other parts of whylogs when we want to know the user's
authentication/configuration status. For example, `log` will check to
see if there is a session and if so, it will use that session to upload
profiles instead of just returning the profiles locally, also printing
out the whylabs url that they can be viewed at. At the moment, `log` is
the only logging api that is aware of sessions. We can follow up with
other apis.

This version is intended to log reference profiles, as opposed to normal
batch profiles that have timestamps. They require an alias along with
their data. The only real distinction between batch profiles and
reference profiles is the downstream use in whylabs and the absence of
time in reference profiles, but they're identical structurally.