Fix to embeddings freq_items and handling of numpy strings (#1170)
Frequent items sketches were empty due to the performance improvement
change to wrapping closest array in numpy. We should check out if
there's still a numpy solution and the effect of this on performance,
but this at least fixes the bug.

This turned up a few other bugs. This PR includes the following:
* numpy subviews are now always ndarrays (there were cases where they
could be pd.Series)
 * unit tests to type check all subviews
 * TypeCountersMetric includes numpy string subview now
 * CardinalityMetric includes numpy string subview now
* FrequentItemsMetric includes numpy string subview (the original bug
that prompted this PR)

---------

Co-authored-by: Richard Rogers <richard@whylabs.ai>
Co-authored-by: richard-rogers <93153899+richard-rogers@users.noreply.github.com>