Add new process logger and thread logger
This PR builds off of the thread based logger that was already in the
experimental name space, which was based on the logger that I had added
to the whylogs container. It essentially ports the container logger out
for use in whylogs as a normal library, and refactored the common parts
out of the thread/process version so they share the same, actor based,
super class.

They process and thread version behave slightly differently and they
have slightly different use cases.

Process Logger:
- Capable of handling profile data for multiple datasets and multiple
  timestamps.
- Runs as a separate process and defers to instances of the thread based
  logger to handle all of the actual logging.
- Enforces a notion of dataset id. Can almost be ignored if the whylabs
  writer isn't the preferred writer, but some id will have to be used.
- Most behavior is inherited from the thread logger as far as logging is
  concerned since it defers entirely to the thread logger.
- Supports synchronous logging if needed, but it has to be explicitly
  enabled since it has more overhead than thread synchronous logging
  because it depends on IPC. It might also interfere with batch
  accumulation which hits performance. This is mostly provided for
  debugging and testing.

Thread Logger:
- Capable of handling multiple timestamps, but only for a single
  dataset id. It doesn't enforce any notion of dataset id.
- Can automatically flush internal state to known writers on a timer.
- Supports synchronous or asynchronous logging, but you'll probably want
  asynchronous.

Using either of them has a few advantages over the current rolling
logger.
- Serialized actions via the actor interface. This means that there
  won't be any concurrency issues that lead to data frame tearing while
  the logger is flushing/writing while the caller is sending messages,
  and it won't block the caller unless the message queue is full.
- Support for historical data. log() takes a timestamp now and the
  loggers will manage the profiles for each timestamp, which includes
  timestamp truncation and other tedious things.
- Bulk processing of messages to minimize internal track() calls.
  whylogs performs way better when you call it fewer times with more
  data. The process logger will handle automatically grouping data and
  merging them into large data frames for more efficient profiling. This
  behavior isn't currently in the thread logger but it can be
  refactored. I held off because the thread logger's performance is
  inherently capped by the gil anyway so the process logger will be
  inevitable for anyone who needs performance.

I think we should retire the current Logger interface now as well. It's
already insufficient for the logging options that we have because the
interface was updated to include properties from each of the loggers
in order to make the types work out. It also uses kwargs for everything
so its very difficult to determine what options even exist without
looking at the source. We should just let people create the logger that
they want with normal python.

This PR can't be pushed until the new session logic is pushed because it
depends on the session to correctly configure the whylabs writer. We're
also going to have a fair bit of documentation to update. Once this is
approved.