Fix whylabs config read error handling (#1201)
## Description

If there is an error reading the server side configuration in WhyLabs
the default value should fallback to an older back-compat format when
uploading profiles to WhyLabs. Also whylogs reads WhyLabs config on each
upload, which if there is a failure results in noisy and slowed down
uploads.

## Changes

- Use older format as fallback when reading configuration fails.
- Only allow using new format if reading config succeeds, and the format
is allowed.
- If there is an error reading whylabs config don't keep attempting to
read that config and repeat this error on subsequent uploads.
- support avoiding reading whylabs config with an env variable

- [x] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).