constraints generator fixes - customizable hll_stddev and error handling (#1242)
## Description

- changes `column_is_probably_unique` to accept `hll_stddev` param
(defaults to 3 - it used to be 1)

This change was motivated by the fact that the `is_probably_unique`
constraint was yielding a high false negative with a `hll_stddev` of 1

- prevent missing metrics from raising an error and instead skip with
warning

When a column has a different type between reference and target
profiles, the auto constraints generator suggests constraints that
requires metrics that are missing in the target profile. That would lead
to an error when calling `add_constraint`. To prevent from crashing,
this PR doesn't raise an error in this case and instead skips the
constraint to be added, along with a warning.


## Changes

- Imperative commit or change title (commit ID) <!-- Add commit ID and
[GitHub will
autolink](https://help.github.com/en/github/writing-on-github/autolinked-references-and-urls).
-->
  - Add some bullet points to explain the change
- Next commit or change (commit ID)

## Related

Relates to organization/repo#number

<!-- Reference related commits, issues and pull requests. Type `#` and
select from the list. -->

Closes organization/repo#number

<!-- List issues for GitHub to automatically close after merge to
default branch. Type `#` and select from the list. See the [GitHub docs
on linking PRs to
issues](https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue)
for more. -->

- [ ] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).