Fix UDF deserialization in column profiles to support UDFs on pyspark (#1271)
## Description

udf metrics are multimetrics but have a different init signature which
leads to problems in deserializing them in pyspark. We need to have a
consistent construction that preserves the udfs submetrics on
deserialization, and works in the pyspark where a column profile is
deserialized from bytes.

## Changes

- support multimetric init shape in udf metric
- add tests to cover missing deserialization path

- [x] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).