change usage stats connection error to info log (#1283)
## Description
The whylogs usage stats emits a log statement at the warning level if it
cannot log the usage statistic, and the message is generic enough it
sounds like an error profiling data. This is confusing at best and not
necessary to log as a warning.

* change the log level to info for this issue, no need for user to take
actions in this case.
* update message to be less confusing

## Related

Fixes #1281

- [x] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).