UDF signature update & Python data types (#1307)
## Description

Changes dataset and type UDF signatures so that they can be coded to
ignore whether they're called with a row or `DataFrame`. Also allows
type-matching `ResolverSpec`s to use Python types rather than whylogs
`DataType`.

## Changes

- Dataset UDF signature is `udf(x: Union[Dict[str, List], pd.DataFrame])
-> Union[List, pd.Series]`
- Type UDF signature is `udf(x: Union[List, pd.Series]) -> Union[List,
pd.Series]`
- UNCHANGED: metric UDF signature remains `udf(x: Any) -> Any`, where
the Anys are single values in a row.
- `UdfSpec` and `ResolverSpec` matching by type can now use native
Python types via a `TypeMapper` in addition to the whylogs `DataType`


- [ ] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).