Several changes to support why.init for authenticated users (#1311)
- Add org id and default dataset id support to sessions.
- Make sessions depend on the new api key format and update the api key
validation to allow for longer api keys that include `:orgId`.
- Remove explicit session type configuration and define a fallback
system for selecting session type based on `allow*` flags, config, and
env state.
- Allow specifying default dataset directly to the session config as an
alternative to env variables, which is how sessions treat api keys.
- Add interactive prompt mode for notebook environments that gets api
key and related config via `input()` and saves it to the new whylogs ini
config.
- Add two new files as pseudo CLIs for setting api key/default model id
in the whylogs config. Without this, the only way to set the config is
via interactive prompt. Supplying that info via env vars doesn't result
in persisting that to the config which would be a strange side effect at
best since the env is presumably still around and definitely taking
priority over the config value.
- Enable interactive log output and url generation for batch profiles,
with or without authentication. This output is disabled in local mode
since no uploading happens there. The output only happens when using the
top level transient logger in `why.log`. The loggers themselves don't
have any output logic at all. It wouldn't make much sense to add output
of this sort to the rolling logger running in the background.
- Add lots of docs to `why.init` since that's the main interface to this
system for users. The goal with these docs is to make it clear what kind
of session you'll be getting since the logic is a little more
complicated than before. The upside to this approach is that most
notebooks, scripts, and production jobs can just have `why.init()` and
not worry about what arguments to supply, so it can truly be ubiquitous.

## Reviewing

To get a sense of the UX introduced here, I recommend running the Guest
Session example notebook, as well as running `python -m
whylogs.api.whylabs.session.why_init` from the command line. The first
one should be pretty representative of a typical user experience after
this change and the second one is a psuedo-cli that users can use to
repeat the first UX without having to spin up a notebook. The command is
similar to running `aws sso configure`, it just gets everything it needs
from a user and stores it locally in the new whylogs.ini config.

Co-authored-by: Anthony Naddeo <anthony@whylabs.ai>