Factor out the whylabs client cache so sessions can use it
The whylabs writer had a ton of complex caching code embedded in it that
depended heavily on self mutations and implicit updates. I tried to
simplify the cache logic and make everything as explicit as I could so
that the client cache would be reused from other places in the
application. This moves the http proxy env var logic into the session
config as well, where all of the config in whylogs now lives.

There were a lot of circular dependencies and initialization race
conditions that made me introduce some lazy initialization. This also
makes the new Session concept mandatory for using whylabs, which should
have minimal impact because we can call it implicitly for people who
don't call it themselves whenever the whylabs writer is used.

This doesn't allow the anonymous sessions to use the private s3
functionality, which seems ok for the moment. The anonymous sessions
don't use the whylabs writer under the hood so that didn't come for
free.

In general, we should always get the latest session with _default_init()
instead of saving a reference to a session because the session can be
reinitialized.