Add graceful behavior when processes are killed
When working with gunicorn I discovered that you'll end up shooting
yourself in the foot if you start the process logger at the wrong time
because gunicorn will end up killing the logger, which results in
apparently nothing happening after you make log calls. This will update
the logger to be explicit about how dead it is and why, throwing on
log() and close() calls.