Switch the pipe signaler to use faster-fifo
This was part of perf optimization work to get batch validation working
faster in the process logger. It needs more work on the validation side
in whylogs in order to fully reap the benefits but its still faster to
use then the older pipe signlaer that was based on mp.Pipe.

The purpose of the signaler is to bridge the asynchronous actor messages
to their sender so they can mimic synchronous invocations. They use a
rendezvous strategy where the sender creates a Future that the pipe
signaler holds and waits for the logger process to signal its
completion.