make to_pandas result same between profiles pre/post uncompounding (#1456)
## Description

Calling `to_pandas()` on an uncompounded profiles will produce a sparse
dataframe. This PR adds a verification step in to_pandas to check
whether the column is the result of uncompounding. If it is, it won't be
included in the resulting DF, producing the same result as the
pre-uncompound profile.

Plus: incidental np.nan bug fix for drift calculation



## Related

Closes organization/repo#number

closes #1401 

- [x] I have reviewed the [Guidelines for Contributing](CONTRIBUTING.md)
and the [Code of Conduct](CODE_OF_CONDUCT.md).

---------

Co-authored-by: felipe207 <felipe@whylabs.ai>