Make process rolling logger optionally org aware
We need to support multi org workflows. Currently, the process rolling
logger stores all of its profiles keyed off of the dataset id, assuming
that the global whylabs api key applies to all of them more or less.
Technically, its already possible to create custom writers which are the
only things that know about the org but the logger would end up rolling
data for different orgs into the same profiles if they have same org id.

This change allows you to explicitly pass an org id into the logger
which is used for cached profiles when logging. That in combination with
a custom writer factory gets you to multi org logging.